test: add tests for Winners
