+++ src/test/kotlin/racingcar/domain/car/CarsTest.kt
@@ -27,20 +27,4 @@ internal class CarsTest {
             assertThat(it.movedDistance).isEqualTo(1)
         }
     }
-
-    @Test
-    internal fun `가장 멀리 간 자동차가 우승자고 우승자는 한 명 이상일 수 있다`() {
-        val expectedWinner1 = Car(name = "가", accelerator = { CarAction.MOVE })
-        val expectedWinner2 = Car(name = "나", accelerator = { CarAction.MOVE })
-        val cars = Cars(
-            cars = listOf(
-                expectedWinner1,
-                expectedWinner2,
-                Car(name = "다", accelerator = { CarAction.STOP }),
-            )
-        )
-        cars.race()
-        assertThat(cars.winners())
-            .containsExactly(expectedWinner1, expectedWinner2)
-    }
 }
